src/producers/requirements.txt
kafka-python>=2.0.2
websockets>=12.0
prometheus-client>=0.19.0
aiohttp>=3.9.0
